requirements.txt
numpy>=1.24.2
opencv_contrib_python>=4.7.0.72
opencv_python>=4.7.0.72
opencv_python_headless>=4.7.0.72
pandas>=1.5.3
plotly>=5.14.1
pytesseract>=0.3.10
PyYAML>=6.0
selenium>=4.8.3
streamlit>=1.21.0
streamlit_authenticator>=0.2.1
lxml>=4.6.2